requirements.txt
django==1.10
django-ajax-selects
django-admin-shortcuts
djangocms_admin_style==0.2.7
django-apptemplates
django-daterange-filter
django-import-export
django-jquery
psycopg2
Click
pyyaml
